requirements.txt
google-cloud-tasks==2.7.0
fastapi==0.70.0